feat: [UF-106] Chip의 타입 분리 및 구조분해, 디자인 수정 #79
