+++ src/components/ui/Chip/Chip.types.ts
@@ -1,0 +1,8 @@
+export interface ChipProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  selected?: boolean;
+  Disabled?: boolean;
+  leftIcon?: React.ReactNode;
+  rightIcon?: React.ReactNode;
+  children: React.ReactNode;
+  dropdown?: React.ReactNode;
+}
